+++ CHANGELOG.md
@@ -8,7 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 Release date `UNRELEASED`
 
-- 
+- Rename `require-config` to `required-config`.
+- Don't use function without `:type` meta as middleware.
+- Allow symbols as middleware type.
+- Add ring core middleware adapters (WIP)
 
 ## `1.0.3-beta1`